Create focus-group-summaries-2019.md
This PR creates a markdown document for the summaries for the 2019 focus groups, and contians the focus group held on the 18th. 

I'd appreciate comments on: 
the formatting
any typos or corrections which need to be made? 

I will add in another pull request the other two summaries, and a contents page and link