Update updated-question-mechanism.md
Hiya! Thanks so much for doing this! It looks great - the glossary is really helpful and we can continue to add to it as we get more feedback. Just made a few corrections to phrasing and layout - if you like the changes, feel free to merge them to your branch and I will merge this into the master. 🌟